Move VmapAdReqeust to AdsModel
